Rename prepare_pool to validate_pool. Now working
